requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysqlclient==1.4.6
protobuf==3.6.1
python-dotenv==0.10.5
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
